# swagger/swagger.yaml
swagger: '2.0'
info:
  description: >-
    Current API provides access to AuthService, which is a part of SecureFiles
    project, which gives you powerfull tool to upload your files in completely
    secure mode and share it with people you trust to see this absolutely secret
    files.
  version: 1.0.2
  title: Auth Service
  termsOfService: 'http://service-name-here/terms/'
  contact:
    email: apiteam@authservice.com
host: 'localhost:8080'
basePath: /v0
produces:
  - application/xml
  - application/json
...
schemes:
  - http
paths:
...
  '/user/files/uploadFile':
    post:
      tags:
        - File
      summary: Uploads a file
      description: Uploads a file, encrypts it and puts to store
      operationId: uploadFile
      consumes:
        - multipart/form-data
      parameters:
        - name: web
          in: header
          description: Header form WebPages (Too long to encrypt/decrypt even small files so we will use ciferservice)
          required: false
          type: string
          format: string
        - name: fileMetadata
          in: body
          description: File metadata to pass to a server
          required: true
          schema:
            $ref: '#/definitions/FileMetadata'
        - name: file
          in: formData
          description: a file to upload
          required: true
          type: file
      responses:
        '201':
          description: a file is uploaded
          schema:
            $ref: '#/definitions/FileMetadata'
        '400':
          description: failed to upload a file
          schema:
            $ref: '#/definitions/Error'
      security:
        - authservice_auth:
            - 'write:files'
            - 'read:files'
...
  '/files/{uuid}':
    get:
      tags:
        - File
      summary: Return a decrypted file
      description: Downloads a file from file storage and returns a decrypted file
      operationId: downloadFile
      produces:
        - application/octet-stream
      parameters:
        - name: desktop
          in: header
          description: Header form Desktop (Too long to decrypt locally)
          required: false
          type: string
          format: string
        - name: uuid
          in: path
          description: file uuid
          required: true
          type: string
      responses:
        '200':
          description: file retrieved
          schema:
            type: file
        '400':
          description: failed to download file
          schema:
            $ref: '#/definitions/Error'
      security:
        - authservice_auth:
            - 'write:files'
            - 'read:files'
...
  '/user/files':
    get:
      tags:
        - File
      summary: List files by filter
      description: This endpoint allows to get list of files searched by filter criteria.
      operationId: listFilesByFilter
      parameters:
        - name: filter
          in: query
          description: object with parameters by which to filter files
          required: true
          schema:
              $ref: '#/definitions/FilesFilter'
      responses:
        '200':
          description: a list of file
          schema:
            type: array
            items:
              $ref: '#/definitions/FileMetadata'
        '400':
          description: failed to retrieve files
          schema:
            $ref: '#/definitions/Error'
      security:
        - authservice_auth:
            - 'write:files'
            - 'read:files'
  '/user/files/{resourceId}/getFileMetadata':
    get:
      tags:
        - File
      summary: Show a file metadata
      description: Shows a file metadata.
      operationId: getFileMetadata
      parameters:
        - name: resourceId
          in: path
          description: ID of a file resource
          required: true
          type: integer
          format: int64
          minimum: 1
      responses:
        '200':
          description: file metadata retrieved
          schema:
            $ref: '#/definitions/FileMetadata'
        '400':
          description: failed to retrieve file metadata
          schema:
            $ref: '#/definitions/Error'
      security:
        - authservice_auth:
            - 'write:files'
            - 'read:files'
  '/user/{userId}/files/{fileId}':
    delete:
      tags:
        - File
      summary: Delete a file
      description: Deletes a file of the user.
      operationId: deleteFile
      parameters:
        - name: userId
          in: path
          description: file owner
          required: true
          type: integer
          format: int64
        - name: fileId
          in: path
          description: file to delete
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: file is deleted
        '400':
          description: failed to delete a file
          schema:
            $ref: '#/definitions/Error'
      security:
        - authservice_auth:
            - 'write:files'
            - 'read:files'
...
securityDefinitions:
  authservice_auth:
    type: oauth2
    authorizationUrl: 'http://projectname.com'
    flow: implicit
    scopes:
      'write:users': modify users
      'read:users': read users
      'write:files': modify files
      'read:files': read files
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      type:
        type: string
      message:
        type: string
    xml:
      name: Error
  FileMetadata:
    type: object
    required:
    - fileName
    - fileSize
    - mime
    - expirationTime
    properties:
      resourceId:
        type: integer
        format: int64
      ownerId:
        type: integer
        format: int64
      fileUUID:
        type: string
      expirationTime:
        type: string
        format: date-time
      fileSize:
        type: integer
        format: int64
      fileName:
        type: string
      mime:
        type: string
      permission:
        $ref: '#/definitions/PermissionType'
      key:
        type: string
    xml:
      name: FileMetadata
...
  PermissionType:
    type: string
    enum:
      - ALL_USERS
      - LIST_OF_USERS
...
  FilesFilter:
    type: object
    required:
    - useId
    - fileType
    properties:
      userId:
        type: integer
        format: int64
        minimum: 1
        description: ID of user to look for files
      offset:
        type: integer
        format: int32
        minimum: 0
        description: Position in pagination
      limit:
        type: integer
        format: int32
        minimum: 5
        maximum: 50
        description:  Number of items to retrieve (5 min, 50 max)
      order:
        type: string
        enum: [ASC, DESC]
        description: Order in which to sort list of files. By defauld sorted by ID
      sortField:
        type: string
        enum: [resourceId, fileName, expirationTime]
        description: Field by which to sort list of files
      permissionType:
        type: string
        enum: [NOT_SET, ALL_USERS, LIST_OF_USERS]
      fileType:
        type: string
        enum: [OWN, SHARED]
        description: Own is for files created by current user, Shared is for files that were shared with current user by other user
    xml:
      name: FilesFilter